+++ packages/mason/lib/src/render.dart
@@ -45,6 +45,9 @@ final _builtInLambdas = <String, LambdaFunction>{
   /// PascalCase
   'pascalCase': (ctx) => ctx.renderString().pascalCase,
 
+  /// Pascal.Dot.Case
+  'pascalDotCase': (ctx) => ctx.renderString().pascalDotCase,
+
   /// param-case
   'paramCase': (ctx) => ctx.renderString().paramCase,