+++ packages/mason_logger/lib/mason_logger.dart
@@ -7,4 +7,10 @@ export 'src/io.dart';
 export 'src/level.dart';
 export 'src/link.dart';
 export 'src/mason_logger.dart'
-    show Logger, Progress, ProgressAnimation, ProgressOptions;
+    show
+        LogStyle,
+        LogTheme,
+        Logger,
+        Progress,
+        ProgressAnimation,
+        ProgressOptions;