+++ packages/mason_logger/lib/src/mason_logger.dart
@@ -6,15 +6,75 @@ import 'package:mason_logger/mason_logger.dart';
 
 part 'progress.dart';
 
+/// Type definition for a function which accepts a log message
+/// and returns a styled version of that message.
+///
+/// Generally, [AnsiCode] values are used to generate a [LogStyle].
+///
+/// ```dart
+/// final alertStyle = (m) => backgroundRed.wrap(styleBold.wrap(white.wrap(m)));
+/// ```
+typedef LogStyle = String? Function(String? message);
+
+String? _detailStyle(String? m) => darkGray.wrap(m);
+String? _infoStyle(String? m) => m;
+String? _errStyle(String? m) => lightRed.wrap(m);
+String? _warnStyle(String? m) => yellow.wrap(styleBold.wrap(m));
+String? _alertStyle(String? m) =>
+    backgroundRed.wrap(styleBold.wrap(white.wrap(m)));
+String? _successStyle(String? m) => lightGreen.wrap(m);
+
+/// {@template log_theme}
+/// A theme object which contains styles for all log message types.
+/// {@endtemplate}
+class LogTheme {
+  /// {@macro log_theme}
+  const LogTheme({
+    LogStyle? detail,
+    LogStyle? info,
+    LogStyle? err,
+    LogStyle? warn,
+    LogStyle? alert,
+    LogStyle? success,
+  })  : detail = detail ?? _detailStyle,
+        info = info ?? _infoStyle,
+        err = err ?? _errStyle,
+        warn = warn ?? _warnStyle,
+        alert = alert ?? _alertStyle,
+        success = success ?? _successStyle;
+
+  /// The [LogStyle] used by [detail].
+  final LogStyle detail;
+
+  /// The [LogStyle] used by [info].
+  final LogStyle info;
+
+  /// The [LogStyle] used by [err].
+  final LogStyle err;
+
+  /// The [LogStyle] used by [warn].
+  final LogStyle warn;
+
+  /// The [LogStyle] used by [alert].
+  final LogStyle alert;
+
+  /// The [LogStyle] used by [success].
+  final LogStyle success;
+}
+
 /// {@template logger}
 /// A basic Logger which wraps `stdio` and applies various styles.
 /// {@endtemplate}
 class Logger {
   /// {@macro logger}
   Logger({
+    this.theme = const LogTheme(),
     this.level = Level.info,
     this.progressOptions = const ProgressOptions(),
   });
+
+  /// The current theme for this logger.
+  final LogTheme theme;
 
   /// The current log level for this logger.
   Level level;
@@ -50,9 +110,10 @@ class Logger {
   void write(String? message) => _stdout.write(message);
 
   /// Writes info message to stdout.
-  void info(String? message) {
+  void info(String? message, {LogStyle? style}) {
     if (level.index > Level.info.index) return;
-    _stdout.writeln(message);
+    style ??= theme.info;
+    _stdout.writeln(style.call(message) ?? message);
   }
 
   /// Writes delayed message to stdout.
@@ -71,34 +132,39 @@ class Logger {
   }
 
   /// Writes error message to stderr.
-  void err(String? message) {
+  void err(String? message, {LogStyle? style}) {
     if (level.index > Level.error.index) return;
-    _stderr.writeln(lightRed.wrap(message));
+    style ??= theme.err;
+    _stderr.writeln(style(message));
   }
 
   /// Writes alert message to stdout.
-  void alert(String? message) {
+  void alert(String? message, {LogStyle? style}) {
     if (level.index > Level.critical.index) return;
-    _stderr.writeln(backgroundRed.wrap(styleBold.wrap(white.wrap(message))));
+    style ??= theme.alert;
+    _stderr.writeln(style(message));
   }
 
   /// Writes detail message to stdout.
-  void detail(String? message) {
+  void detail(String? message, {LogStyle? style}) {
     if (level.index > Level.debug.index) return;
-    _stdout.writeln(darkGray.wrap(message));
+    style ??= theme.detail;
+    _stdout.writeln(style(message));
   }
 
   /// Writes warning message to stderr.
-  void warn(String? message, {String tag = 'WARN'}) {
+  void warn(String? message, {String tag = 'WARN', LogStyle? style}) {
     if (level.index > Level.warning.index) return;
     final output = tag.isEmpty ? '$message' : '[$tag] $message';
-    _stderr.writeln(yellow.wrap(styleBold.wrap(output)));
+    style ??= theme.warn;
+    _stderr.writeln(style(output));
   }
 
   /// Writes success message to stdout.
-  void success(String? message) {
+  void success(String? message, {LogStyle? style}) {
     if (level.index > Level.info.index) return;
-    _stdout.writeln(lightGreen.wrap(message));
+    style ??= theme.success;
+    _stdout.writeln(style(message));
   }
 
   /// Prompts user and returns response.
@@ -315,7 +381,11 @@ class Logger {
           // show cursor
           ..write('\x1b[?25h')
           ..write('$message ')
-          ..writeln(styleDim.wrap(lightCyan.wrap('$results')));
+          ..writeln(
+            styleDim.wrap(
+              lightCyan.wrap('${results.map(resolvedDisplay).toList()}'),
+            ),
+          );
 
         break;
       }